change type from int to long
git-tfs-id: [http://developmentvlan:8585/tfs/vanrise.collection]$/;C37997

+++ Retail/Code/Retail.Invoice.Business/Extensions/RetailSubscriberPartnerSettings.cs
@@ -116,7 +116,7 @@ namespace Retail.Invoice.Business
         public override string GetPartnerName(IPartnerNameManagerContext context)
         {
             AccountManager accountManager = new AccountManager();
-            return accountManager.GetAccountName(Convert.ToInt32(context.PartnerId));
+            return accountManager.GetAccountName(Convert.ToInt64(context.PartnerId));
         }
 
         private void AddRDLCParameter(Dictionary<string, VRRdlcReportParameter> rdlcReportParameters, RDLCParameter key, string value, bool isVisible)
